requirements.txt
ultralytics>=8.0.0
streamlit>=1.24.0
opencv-python>=4.7.0
numpy>=1.24.0
pillow>=9.4.0
roboflow>=1.0.5
matplotlib>=3.7.0
pandas>=1.5.3
altair>=4.2.2
gitpython>=3.1.30